pulled chat container logic out of modal file to separate duties
